+++ src/transformer/metadata-decorator.ts
@@ -2,10 +2,6 @@ import ts from 'typescript';
 import { serialize } from './serialize';
 
 export function metadataDecorator(key : string, object : any) {
-    // let metadataFuncExpr : ts.Expression = ts.factory.createPropertyAccessExpression(
-    //     ts.factory.createIdentifier('Reflect'),
-    //     ts.factory.createIdentifier('metadata')
-    // );
     let metadataFuncExpr : ts.Expression = ts.factory.createIdentifier('__RtΦ');
 
     return ts.factory.createDecorator(
@@ -16,5 +12,21 @@ export function metadataDecorator(key : string, object : any) {
                 serialize(object)
             ]
         )
+    )
+}
+
+export function directMetadataDecorator(key : string, object : any) {
+    let dec = metadataDecorator(key, object)
+    dec['__Φdirect'] = true;
+    return dec;
+}
+
+export function decorateFunctionExpression(func : ts.FunctionExpression, decorators : ts.Decorator[]) {
+    return ts.factory.createCallExpression(
+        ts.factory.createIdentifier('__RfΦ'),
+        [], [
+            func,
+            ts.factory.createArrayLiteralExpression(decorators.map(d => d.expression))
+        ]
     )
 }