+++ src/transformer/rt-helper.ts
@@ -51,3 +51,75 @@ export function rtHelper() {
         )
     );
 }
+
+export function rtfHelper() {
+    const factory = ts.factory;
+    return factory.createVariableStatement(
+        undefined,
+        factory.createVariableDeclarationList(
+            [factory.createVariableDeclaration(
+                factory.createIdentifier("__RfΦ"),
+                undefined,
+                undefined,
+                factory.createArrowFunction(
+                    undefined,
+                    undefined,
+                    [
+                        factory.createParameterDeclaration(
+                            undefined,
+                            undefined,
+                            undefined,
+                            factory.createIdentifier("f"),
+                            undefined,
+                            undefined,
+                            undefined
+                        ),
+                        factory.createParameterDeclaration(
+                            undefined,
+                            undefined,
+                            undefined,
+                            factory.createIdentifier("d"),
+                            undefined,
+                            undefined,
+                            undefined
+                        )
+                    ],
+                    undefined,
+                    factory.createToken(ts.SyntaxKind.EqualsGreaterThanToken),
+                    factory.createParenthesizedExpression(factory.createBinaryExpression(
+                        factory.createCallExpression(
+                            factory.createPropertyAccessExpression(
+                                factory.createIdentifier("d"),
+                                factory.createIdentifier("forEach")
+                            ),
+                            undefined,
+                            [factory.createArrowFunction(
+                                undefined,
+                                undefined,
+                                [factory.createParameterDeclaration(
+                                    undefined,
+                                    undefined,
+                                    undefined,
+                                    factory.createIdentifier("d"),
+                                    undefined,
+                                    undefined,
+                                    undefined
+                                )],
+                                undefined,
+                                factory.createToken(ts.SyntaxKind.EqualsGreaterThanToken),
+                                factory.createCallExpression(
+                                    factory.createIdentifier("d"),
+                                    undefined,
+                                    [factory.createIdentifier("f")]
+                                )
+                            )]
+                        ),
+                        factory.createToken(ts.SyntaxKind.CommaToken),
+                        factory.createIdentifier("f")
+                    ))
+                )
+            )],
+            ts.NodeFlags.Const
+        )
+    )
+}